button micro-interactions, color setting, and submission handle adjustment
setup button disabled animation and accessible typography colors for form
