+++ app/components/atoms/Button/ButtonProps.tsx
@@ -10,31 +10,36 @@ export type ButtonProps = {
 	 * The variant of the button.
 	 * @default 'primary'
 	 */
-	variant?: 'primary' | 'secondary' | 'danger'
+	variant?: 'primary' | 'secondary' | 'danger';
+
+	/**
+	 * Specifies whether the button is disabled.
+	 */
+	disabled?: boolean;
 
 	/**
 	 * The type of the button.
 	 * @default 'button'
 	 */
-	type?: 'button' | 'submit' | 'reset'
+	type?: 'button' | 'submit' | 'reset';
 
 	/**
 	 * The icon to be displayed on the button.
 	 */
-	icon?: IconDefinition
+	icon?: IconDefinition;
 
 	/**
 	 * The content to be displayed inside the button.
 	 */
-	children: ReactNode
+	children: ReactNode;
 
 	/**
 	 * The function to be called when the button is clicked.
 	 */
-	onClick?: () => void
+	onClick?: () => void;
 
 	/**
 	 * The CSS class name for the button.
 	 */
-	className?: string
-}
+	className?: string;
+};
