+++ package.json
@@ -75,7 +75,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "remix-utils": "^7.6.0",
-    "remix-validated-form": "^5.1.5"
+    "remix-validated-form": "^5.1.5",
+    "zod-form-data": "^2.0.2"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.0.2",
